First step towards time feature
Wrote an example game that uses a new TimeEngine physics engine that can update the color shade of every sprite in the game on each update.

Still need to make the colors match change in time, either by importing the clock or recording the passage of time in the game object.