Improved relative import mechanism.
We need some way of locating game assets (images, sounds, ink files,
etc.) that are packaged with Quest, so that example games can be played
from anywhere on the filesystem. Previously, we had a function called
resolve_path in each example file, which located assets relative to that
file. This function had to be defined in each file, because it relied on
__file__ to refer to the file itself (not somewhere else it got imported
from). This was less than ideal because the examples need to be as
minimal and simple as possible, and they started off with this annoying
detail.

I implemented a better solution, using helpers.resolve_resource_path,
which locates assets relative to the root of the installed quest
package. Now this helper can be imported into the examples instead of
defined in each file. I updated the examples to use this file path
resolution strategy.